legacy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iceberg-janitor"
version = "0.1.0"
description = "K8s-native Iceberg table maintenance — catalog-only, no direct S3 access"
readme = {text = "K8s-native Iceberg table maintenance system", content-type = "text/plain"}
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "pyiceberg[s3,glue]>=0.7.0",
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "prometheus-client>=0.20",
    "pyyaml>=6.0",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.110", "uvicorn>=0.29", "cloudevents>=1.10"]
query = ["duckdb>=1.1.0"]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy"]

[project.scripts]
janitor = "iceberg_janitor.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/iceberg_janitor"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
